[Refactor] 모듈화
