+++ src/components/ShareButton/ShareButton.jsx
@@ -19,7 +19,7 @@ export default function ShareButton() {
   // 카카오 SDK 초기화
   useEffect(() => {
     const initializeKakao = () => {
-      window.Kakao.init(JAVASCRIPT_KEY);
+      window.Kakao.init("JAVASCRIPT_KEY");
     };
 
     if (document.readyState === "complete") {
